refactor: update examples
