+++ examples/awesome-vault-service/src/bin/challenger.rs
@@ -3,7 +3,7 @@
 use alloy::primitives::Address;
 use awesome_vault_service::{
     bindings::awesomevaulttaskmanager::AwesomeVaultTaskManager::AwesomeVaultTaskManagerInstance,
-    task_manager::save_value,
+    response_calculator::VaultServiceResponseCalculator,
 };
 use eigensdk::{
     challenger::{
@@ -29,10 +29,11 @@ async fn main() -> Result<()> {
 
     let contract = AwesomeVaultTaskManagerInstance::new(task_manager_address, wallet);
 
-    let redis_state = Arc::new(Mutex::new(BTreeMap::new()));
-    let compute = save_value(redis_state.clone()).await;
+    let vault_service_response_calculator = VaultServiceResponseCalculator {
+        vault: Arc::new(Mutex::new(BTreeMap::new())),
+    };
 
-    let is_response_correct = verifier_from_compute_function(compute);
+    let is_response_correct = verifier_from_compute_function(vault_service_response_calculator);
     let task_processor = IndexingChallengerProcessor::new(contract, is_response_correct);
     let config = ChallengerConfig {
         http_rpc_url,
