+++ examples/awesome-vault-service/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod bindings;
+pub mod response_calculator;
 pub mod task_manager;
 pub mod utils;
